Frontend: Move subroutine parser contracts into frontend suites
