+++ loki/tests/test_subroutine.py
@@ -10,9 +10,9 @@ from pathlib import Path
 import pytest
 
 from loki import Sourcefile, Module, Subroutine, Function, fgen, fexprgen
-from loki.jit_build import jit_compile, jit_compile_lib, clean_test
+from loki.jit_build import jit_compile, clean_test
 from loki.expression import symbols as sym
-from loki.frontend import available_frontends, OMNI, REGEX
+from loki.frontend import available_frontends, OMNI
 from loki.ir import (
     nodes as ir, FindNodes, FindVariables, FindTypedSymbols,
     Transformer
@@ -930,172 +930,6 @@ end subroutine routine_empty_spec
     assert len(routine.body) == 1
 
 
-@pytest.mark.parametrize('frontend', available_frontends())
-def test_external_stmt(tmp_path, frontend):
-    """
-    Tests procedures passed as dummy arguments and declared as EXTERNAL.
-    """
-    fcode_external = """
-! This should be tested as well with interface statements in the caller
-! routine, and the subprogram definitions outside (to have "truly external"
-! procedures, however, we need to make the INTERFACE support more robust first
-
-subroutine other_external_subroutine(outvar)
-  implicit none
-  integer, intent(out) :: outvar
-  outvar = 4
-end subroutine other_external_subroutine
-
-function other_external_function() result(outvar)
-  implicit none
-  integer :: outvar
-  outvar = 6
-end function other_external_function
-    """.strip()
-
-    fcode = """
-subroutine routine_external_stmt(invar, sub1, sub2, sub3, outvar, func1, func2, func3)
-  implicit none
-  integer, intent(in) :: invar
-  external sub1
-  external :: sub2, sub3
-  integer, intent(out) :: outvar
-  integer, external :: func1, func2
-  integer, external :: func3
-  integer tmp
-
-  call sub1(tmp)
-  outvar = invar + tmp  ! invar + 1
-  call sub2(tmp)
-  outvar = outvar + tmp + func1()  ! (invar + 1) + 1 + 6
-  call sub3(tmp)
-  outvar = outvar + tmp + func2()  ! (invar + 8) + 4 + 2
-  tmp = func3()
-  outvar = outvar + tmp  ! (invar + 14) + 2
-end subroutine routine_external_stmt
-
-subroutine routine_call_external_stmt(invar, outvar)
-  implicit none
-  integer, intent(in) :: invar
-  integer, intent(out) :: outvar
-
-  interface
-    subroutine other_external_subroutine(outvar)
-      integer, intent(out) :: outvar
-    end subroutine other_external_subroutine
-  end interface
-
-  interface
-    function other_external_function()
-      integer :: other_external_function
-    end function other_external_function
-  end interface
-
-  call routine_external_stmt(invar, external_subroutine, external_subroutine, other_external_subroutine, &
-                            &outvar, other_external_function, external_function, external_function)
-
-contains
-
-  subroutine external_subroutine(outvar)
-    implicit none
-    integer, intent(out) :: outvar
-    outvar = 1
-  end subroutine external_subroutine
-
-  function external_function()
-    implicit none
-    integer :: external_function
-    external_function = 2
-  end function external_function
-
-end subroutine routine_call_external_stmt
-    """.strip()
-
-    source = Sourcefile.from_source(fcode, frontend=frontend)
-    routine = source['routine_external_stmt']
-    assert len(routine.arguments) == 8
-
-    for decl in FindNodes(ir.ProcedureDeclaration).visit(routine.spec):
-        # Is the EXTERNAL attribute set?
-        assert decl.external
-        for v in decl.symbols:
-            # Are procedure names represented as Scalar objects?
-            assert isinstance(v, sym.ProcedureSymbol)
-            assert isinstance(v.type.dtype, ProcedureType)
-            assert v.type.external is True
-            assert v.type.dtype.procedure == BasicType.DEFERRED
-            if 'sub' in v.name:
-                assert not v.type.dtype.is_function
-                assert v.type.dtype.return_type is None
-            else:
-                assert v.type.dtype.is_function
-                assert v.type.dtype.return_type.compare(SymbolAttributes(BasicType.INTEGER))
-
-    # Generate code, compile and load
-    extpath = tmp_path/(f'subroutine_routine_external_{frontend}.f90')
-    with extpath.open('w') as f:
-        f.write(fcode_external)
-    filepath = tmp_path/(f'subroutine_routine_external_stmt_{frontend}.f90')
-    source.path = filepath
-    lib = jit_compile_lib([source, extpath], path=tmp_path, name='subroutine_external')
-    function = lib.routine_call_external_stmt
-
-    outvar = function(7)
-    assert outvar == 23
-    clean_test(filepath)
-
-
-@pytest.mark.parametrize('frontend', available_frontends())
-def test_subroutine_interface(tmp_path, frontend, header_path):
-    """
-    Test auto-generation of an interface block for a given subroutine.
-    """
-    fcode = """
-subroutine test_subroutine_interface (in1, in2, in3, out1, out2)
-  use header, only: jprb
-  IMPLICIT NONE
-  integer, intent(in) :: in1, in2
-  real(kind=jprb), intent(in) :: in3(in1, in2)
-  real(kind=jprb), intent(out) :: out1, out2
-  integer :: localvar
-  localvar = in1 + in2
-  out1 = real(localvar, kind=jprb)
-  out2 = out1 + 2.
-end subroutine
-"""
-    if frontend == OMNI:
-        # Generate xmod
-        Sourcefile.from_file(header_path, frontend=frontend, xmods=[tmp_path])
-
-    routine = Subroutine.from_source(fcode, xmods=[tmp_path], frontend=frontend)
-
-    if frontend == OMNI:
-        assert fgen(routine.interface).strip() == """
-INTERFACE
-  SUBROUTINE test_subroutine_interface (in1, in2, in3, out1, out2)
-    USE header, ONLY: jprb
-    IMPLICIT NONE
-    INTEGER, INTENT(IN) :: in1
-    INTEGER, INTENT(IN) :: in2
-    REAL(KIND=selected_real_kind(13, 300)), INTENT(IN) :: in3(in1, in2)
-    REAL(KIND=selected_real_kind(13, 300)), INTENT(OUT) :: out1
-    REAL(KIND=selected_real_kind(13, 300)), INTENT(OUT) :: out2
-  END SUBROUTINE test_subroutine_interface
-END INTERFACE
-""".strip()
-    else:
-        assert fgen(routine.interface).strip() == """
-INTERFACE
-  SUBROUTINE test_subroutine_interface (in1, in2, in3, out1, out2)
-    USE header, ONLY: jprb
-    IMPLICIT NONE
-    INTEGER, INTENT(IN) :: in1, in2
-    REAL(KIND=jprb), INTENT(IN) :: in3(in1, in2)
-    REAL(KIND=jprb), INTENT(OUT) :: out1, out2
-  END SUBROUTINE test_subroutine_interface
-END INTERFACE
-""".strip()
-
 
 @pytest.mark.parametrize('frontend', available_frontends())
 def test_subroutine_rescope_symbols(tmp_path, frontend):
@@ -1387,26 +1221,6 @@ end subroutine subroutine_stmt_func
     clean_test(filepath)
 
 
-@pytest.mark.parametrize('frontend', available_frontends())
-def test_mixed_declaration_interface(frontend):
-    """
-    A simple test to catch and shame mixed declarations.
-    """
-    fcode = """
-subroutine valid_fortran(i, m)
-   integer :: i, j, m
-   integer :: k,l
-end subroutine valid_fortran
-"""
-
-    with pytest.raises(AssertionError) as error:
-        routine = Subroutine.from_source(fcode, frontend=frontend)
-        assert isinstance(routine.body, ir.Section)
-        assert isinstance(routine.spec, ir.Section)
-        _ = routine.interface
-
-    assert "Declarations must have intents" in str(error.value)
-
 
 @pytest.mark.parametrize('frontend', available_frontends())
 def test_subroutine_comparison(frontend):
@@ -1483,134 +1297,6 @@ end subroutine my_routine
     assert not r1 == r2
 
 
-@pytest.mark.parametrize('frontend', available_frontends())
-def test_subroutine_lazy_arguments_incomplete1(frontend):
-    """
-    Test that argument lists for subroutines are correctly captured when the object is made
-    complete.
-
-    The rationale for this test is that for dummy argument lists with interleaved comments and line
-    breaks, matching is non-trivial and, since we don't currently need the argument list
-    in the incomplete REGEX-parsed IR, we accept that this information is incomplete initially.
-    tmp_path, we make sure this information is captured correctly after completing the full frontend
-    parse.
-    """
-    fcode = """
-subroutine my_routine(n, a, b, d)
-    integer, intent(in) :: n
-    real, intent(in) :: a(n), b(n)
-    real, intent(out) :: d(n)
-    integer :: i
-
-    do i=1, n
-        d(i) = a(i) + b(i)
-    end do
-end subroutine my_routine
-    """.strip()
-
-    routine = Subroutine.from_source(fcode, frontend=REGEX)
-    assert routine._incomplete
-    assert routine.arguments == ()
-    assert routine.argnames == []
-    assert routine._dummies == ()
-    assert all(isinstance(arg, sym.DeferredTypeSymbol) for arg in routine.arguments)
-
-    routine.make_complete(frontend=frontend)
-    assert not routine._incomplete
-    assert routine.arguments == ('n', 'a(n)', 'b(n)', 'd(n)')
-    assert routine.argnames == ['n', 'a', 'b', 'd']
-    assert routine._dummies == ('n', 'a', 'b', 'd')
-    assert isinstance(routine.arguments[0], sym.Scalar)
-    assert all(isinstance(arg, sym.Array) for arg in routine.arguments[1:])
-
-
-@pytest.mark.parametrize('frontend', available_frontends())
-def test_subroutine_lazy_arguments_incomplete2(frontend):
-    """
-    Test that argument lists for subroutines are correctly captured when the object is made
-    complete.
-
-    The rationale for this test is that for dummy argument lists with interleaved comments and line
-    breaks, matching is non-trivial and, since we don't currently need the argument list
-    in the incomplete REGEX-parsed IR, we accept that this information is not available initially.
-    tmp_path, we make sure this information is captured correctly after completing the full frontend
-    parse.
-    """
-    fcode = """
-SUBROUTINE CLOUDSC &
- !---input
- & (KIDIA,    KFDIA,    KLON,    KLEV,&
- & PT, PQ, &
- !---prognostic fields
- & PA,&
- & PCLV,  &
- & PSUPSAT,&
-!-- arrays for aerosol-cloud interactions
-!!! & PQAER,    KAER, &
- & PRE_ICE,&
- & PCCN,     PNICE,&
- !---diagnostic output
- & PCOVPTOT, PRAINFRAC_TOPRFZ,&
- !---resulting fluxes
- & PFSQLF,   PFSQIF ,  PFCQNNG,  PFCQLNG&
- & )
-IMPLICIT NONE
-INTEGER, PARAMETER :: JPIM = SELECTED_INT_KIND(9)
-INTEGER, PARAMETER :: JPRB = SELECTED_REAL_KIND(13,300)
-INTEGER(KIND=JPIM),PARAMETER :: NCLV=5      ! number of microphysics variables
-INTEGER(KIND=JPIM),INTENT(IN)    :: KLON             ! Number of grid points
-INTEGER(KIND=JPIM),INTENT(IN)    :: KLEV             ! Number of levels
-INTEGER(KIND=JPIM),INTENT(IN)    :: KIDIA
-INTEGER(KIND=JPIM),INTENT(IN)    :: KFDIA
-REAL(KIND=JPRB)   ,INTENT(IN)    :: PT(KLON,KLEV)    ! T at start of callpar
-REAL(KIND=JPRB)   ,INTENT(IN)    :: PQ(KLON,KLEV)    ! Q at start of callpar
-REAL(KIND=JPRB)   ,INTENT(IN)    :: PA(KLON,KLEV)    ! Original Cloud fraction (t)
-REAL(KIND=JPRB)   ,INTENT(IN)    :: PCLV(KLON,KLEV,NCLV)
-REAL(KIND=JPRB)   ,INTENT(IN)    :: PSUPSAT(KLON,KLEV)
-REAL(KIND=JPRB)   ,INTENT(IN)    :: PRE_ICE(KLON,KLEV)
-REAL(KIND=JPRB)   ,INTENT(IN)    :: PCCN(KLON,KLEV)     ! liquid cloud condensation nuclei
-REAL(KIND=JPRB)   ,INTENT(IN)    :: PNICE(KLON,KLEV)    ! ice number concentration (cf. CCN)
-REAL(KIND=JPRB)   ,INTENT(OUT)   :: PCOVPTOT(KLON,KLEV) ! Precip fraction
-REAL(KIND=JPRB)   ,INTENT(OUT)   :: PRAINFRAC_TOPRFZ(KLON)
-REAL(KIND=JPRB)   ,INTENT(OUT)   :: PFSQLF(KLON,KLEV+1)  ! Flux of liquid
-REAL(KIND=JPRB)   ,INTENT(OUT)   :: PFSQIF(KLON,KLEV+1)  ! Flux of ice
-REAL(KIND=JPRB)   ,INTENT(OUT)   :: PFCQLNG(KLON,KLEV+1) ! -ve corr for liq
-REAL(KIND=JPRB)   ,INTENT(OUT)   :: PFCQNNG(KLON,KLEV+1) ! -ve corr for ice
-END SUBROUTINE CLOUDSC
-    """.strip()
-
-    argnames = (
-        'kidia', 'kfdia', 'klon', 'klev', 'pt', 'pq',
-        'pa', 'pclv', 'psupsat',
-        'pre_ice', 'pccn', 'pnice',
-        'pcovptot', 'prainfrac_toprfz',
-        'pfsqlf', 'pfsqif', 'pfcqnng', 'pfcqlng'
-    )
-    argnames_with_dim = (
-        'kidia', 'kfdia', 'klon', 'klev', 'pt(klon, klev)', 'pq(klon, klev)',
-        'pa(klon, klev)', 'pclv(klon, klev, nclv)', 'psupsat(klon, klev)',
-        'pre_ice(klon, klev)', 'pccn(klon, klev)', 'pnice(klon, klev)',
-        'pcovptot(klon, klev)', 'prainfrac_toprfz(klon)',
-        'pfsqlf(klon, klev + 1)', 'pfsqif(klon, klev + 1)', 'pfcqnng(klon, klev + 1)', 'pfcqlng(klon, klev + 1)'
-    )
-
-    routine = Subroutine.from_source(fcode, frontend=REGEX)
-    assert routine._incomplete
-    # NOTE: This represents the current capabilities of the REGEX frontend. If this test
-    # suddenly fails because the argument list happens to be captured correctly:
-    # Nice one! Go ahead and change the test.
-    assert routine.arguments == ()
-    assert routine.argnames == []
-    assert routine._dummies == ()
-    assert all(isinstance(arg, sym.DeferredTypeSymbol) for arg in routine.arguments)
-
-    routine.make_complete(frontend=frontend)
-    assert not routine._incomplete
-    assert routine.arguments == argnames_with_dim
-    assert [arg.upper() for arg in routine.argnames] == [arg.upper() for arg in argnames]
-    assert routine._dummies == argnames
-    assert all(isinstance(arg, sym.Scalar) for arg in routine.arguments[:4])
-    assert all(isinstance(arg, sym.Array) for arg in routine.arguments[4:])
 
 
 @pytest.mark.parametrize('frontend', available_frontends())
